requirements.txt
# Video redaction dependencies
opencv-python
presidio-analyzer
ipykernel
transformers
datasets
protobuf
python-doctr
Levenshtein
accelerate

# Audio redaction dependencies
faster-whisper>=1.0.0
librosa>=0.9.0
ffmpeg-python>=0.2.0
google-genai

# Shared dependencies
streamlit>=1.45.0
python-dotenv>=1.0.0

# Development and debugging
setuptools>=65.0.0